Fix output and Makefile

+++ my_calc/lib/include/functional.h
@@ -4,6 +4,6 @@
 int addition(const int a, const int b);
 int substraction(const int a, const int b);
 int multiplication(const int a, const int b);
-double division(const int a, const int b);
+double division(const double a, const double b);
 
 #endif
